pyproject.toml
[project]
name = "chat-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Ankaj",email = "42106362+coderankaj@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "channels[daphne] (>=4.2.0,<5.0.0)",
    "channels-redis (>=4.2.1,<5.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "daphne (>=4.1.2,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
